--- debug.py
@@ -1,163 +1,0 @@
-from dotenv import load_dotenv
-from langchain_openai import ChatOpenAI
-from loaders.vectorstore import *
-import json
-import re
-import glob
-
-
-class SummaryBot:
-    def __init__(self, model_name, streaming, temperature, vectorstore):
-        self.llm = ChatOpenAI(
-            model_name=model_name, streaming=streaming, temperature=temperature
-        )
-        self.vectorstore = vectorstore
-
-    def summarize(self, company):
-        special_terms = self.vectorstore.loader.special_terms
-        special_terms_name_list = [term["name"] for term in special_terms]
-        insurance_info = {"company": company, "insurance": get_insurance(company)}
-        special_terms_info = []
-
-        for term_name in special_terms_name_list:
-            term_summary = self.retrieve_term_summary(term_name)
-            illness = self.infer_illness(term_name, term_summary)
-            info = {
-                "name": clean_text(term_name),
-                "summary": term_summary,
-                "illness": illness,
-            }
-            special_terms_info.append(info)
-
-        insurance_info["special_terms"] = special_terms_info
-        return insurance_info
-
-    def retrieve_term_summary(self, term_name):
-        prompt = f"""
-        아래 특약 이름에 대한 요약 정보를 작성하세요:
-        특약 이름: {term_name}
-        출력 형식:
-        {{
-            "details": {{
-                "causes": "보험금 지급 사유",
-                "limits": "보장 금액 한도",
-                "details": "특약 요약"
-            }}
-        }}
-        정보가 부족한 경우라도 지정된 형식에 따라 적절한 정보를 생성하세요.
-        """
-        response = self.llm(prompt)
-
-        try:
-            # JSON으로 응답 파싱
-            details_json = json.loads(response.content)
-            # 'details' 키가 있는 경우 반환, 없으면 기본값 제공
-            return details_json.get(
-                "details",
-                {
-                    "causes": "보험금 지급 사유 없음",
-                    "limits": "보장 금액 한도 없음",
-                    "details": "특약 요약 없음",
-                },
-            )
-        except json.JSONDecodeError:
-            # JSON 파싱 실패 시 기본값 반환
-            return {
-                "causes": "보험금 지급 사유 없음",
-                "limits": "보장 금액 한도 없음",
-                "details": "특약 요약 없음",
-            }
-
-    def infer_illness(self, term_name, details):
-        prompt = f"""
-        아래 특약 이름과 설명에 따라 관련된 질병을 유추하세요:
-        특약 이름: {term_name}
-        특약 설명: {details}
-        가능한 질병 카테고리: [백내장, 슬관절, 치과, 상해, 사망, MRI, 약물, 재활, 피부]
-        출력 형식:
-        {{
-            "illness": 관련된 질병 카테고리 (하나 또는 여러 개 가능)를 list 로 반환
-        }}
-        """
-        response = self.llm(prompt)
-        try:
-            illness_json = json.loads(response.content)
-            return illness_json.get("illness", "기타")
-        except json.JSONDecodeError:
-            return "기타"
-
-
-def get_insurance(company):
-    insurance_items = {
-        "DB_cat": "무배당 다이렉트 펫블리 반려묘보험",
-        "DB_dog": "무배당 다이렉트 펫블리 반려견보험",
-        "hyundai_dog": "무배당 현대해상다이렉트굿앤굿 우리펫보험",
-        "hyundai_cat": "무배당 현대해상다이렉트굿앤굿 우리펫보험",
-        "KB_dog": "KB 다이렉트 금쪽같은 펫보험 (강아지) (무배당)",
-        "KB_cat": "KB 다이렉트 금쪽같은 펫보험 (고양이) (무배당)",
-        "meritz_dog": "(무)펫퍼민트 Puppy&Family 보험 다이렉트",
-        "meritz_cat": "(무)펫퍼민트 Cat&Family 보험 다이렉트",
-        "samsung_dog": "무배당 삼성화재 다이렉트 반려견보험",
-        "samsung_cat": "무배당 삼성화재 다이렉트 반려묘보험",
-    }
-    return insurance_items.get(company)
-
-
-def clean_json(text):
-    cleaned_text = text.replace("\n", "").replace("    ", "").strip()
-
-    try:
-        cleaned_text = f"[{cleaned_text.replace('}{', '},{')}]"
-        return json.loads(cleaned_text)
-    except json.JSONDecodeError as e:
-        print(f"Error parsing JSON: {e}")
-        return {}
-
-
-def clean_text(text):
-    filtered_text = re.sub(r"[^가-힣0-9().]", " ", text)
-    filtered_text = re.sub(r"\(\s*\)", "", filtered_text)
-    cleaned_text = re.sub(r"\s+", " ", filtered_text).strip()
-
-    return cleaned_text
-
-
-def save_summaries(company, form):
-    loader_path = f"data/dataloaders/{company}_loader.pkl"
-    loader = load_loader(loader_path)
-    vectordb = load_vectorstore(f"{company}_store", loader)
-
-    bot = SummaryBot(
-        model_name="gpt-4o-mini", streaming=False, temperature=0, vectorstore=vectordb
-    )
-
-    summary = bot.summarize(company)
-
-    pet_type = company.split("_")[1]
-    output_filename = f"summaries/{pet_type}/{company}_{form}.json"
-    with open(output_filename, "w", encoding="utf-8") as f:
-        json.dump(summary, f, ensure_ascii=False, indent=4)
-
-
-if __name__ == "__main__":
-    load_dotenv()
-
-    file_paths = glob.glob(f"data/pdf/*.pdf")
-
-    for path in file_paths:
-        company = extract_company_name(path)
-        save_summaries(company, "summary")
-        print(f"summary for {path} saved")
-
-    # ---------debug by file------------
-    # company = "DB_dog"
-    # loader_path = f"data/dataloaders/{company}_loader.pkl"
-    # loader = load_loader(loader_path)
-    # vectordb = load_vectorstore(f"{company}_store", loader)
-
-    # bot = SummaryBot(
-    #     model_name="gpt-4o-mini", streaming=False, temperature=0.3, vectorstore=vectordb
-    # )
-
-    # summary = bot.summarize(company)
-    # print(summary)
